Avoid compiler warning.

+++ backend/new_cxx/src/pytrace.cxx
@@ -175,7 +175,7 @@ struct Inferer
   {
     if (trace->numUnconstrainedChoices() == 0) { return; }
     
-    for (int i = 0; i < transitions; ++i)
+    for (size_t i = 0; i < transitions; ++i)
     {
       mixMH(trace.get(), scaffoldIndexer, gKernel);
